Finally, code that will train
The DDPG model critic archetecture was a deviation from what was provided
in class.  After comparing my code to @tommytracey, that was the only
funamental difference.  Based on a sample size of 15, it now appears to
work.  Also, this code is set up to do a grid search, which may go away in
the future.